streamlit/GeminiMind/pyproject.toml
[tool.poetry]
name = "geminimind"
version = "0.1.0"
description = ""
authors = ["Mubashar-Bashir <123990120+Mubashar-Bashir@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
google-generativeai = "^0.8.1"
python-dotenv = "^1.0.1"
markdown = "^3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
